Make variants matching algorithm more generic (fixes Chez Luc stairs replacement issue)
